Now returns a rectified neural vector.
As the BOLD response (roughly) is sensitive to total synatptic activity, as opposed to inhibitory or excitatory responses, the neural vector is rectified prior to being returned.

We may wish to return to this and examine rectification as a parameter that can be adjusted. There are conflicting findings in the literature relating absolute or rectified LFP to BOLD response.

Signed-off-by: gkaguirre <gkaguirre@me.com>