BTRM toolbox
Blocked temporal response model.
Derivied from Ben’s code in the BDCM. The goal is to divide the codebase between the portion that falls within the general temporal fitting engine (and thus is part of the BTRM toolbox) and the part that idiosyncratic to the Mt. Sinai 7T dataset, which will live in an Analysis repository called mriBlockFrequencyDirectionAnalysis. Much hacking ahead.

Signed-off-by: gkaguirre <gkaguirre@me.com>