Added verbosity key/value pair to parent class
You should now be able to access field  obj.verbosity in any method to get a string with the level of verbosity passed to the object constructor.  Current values are ‘none’ (default) and ‘high’ (dump as much as possible).

Implementing this involved editing the constructor for tfe and then for all the subclasses so that they call the parent class constructor.  I did this for all classes except BDCM, which I renamed with BecomingDeprecated.  BDCM won’t run correctly anymore because of this change.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>