Removed the aggregateParams method
This was originally created to be used for crossValidation routines. It became clear, however, that it would not actually be useful for this purpose because the combination of paramFit results must know about the particular stimulus types to be able to appropriately combine paramFit results. Therefore, this functionality is not implemented as a top-level method, but instead is wrapped into the CrossValidation function itself.

Signed-off-by: gkaguirre <gkaguirre@me.com>