Replace hash-based regression check with a more robust check
The problem with using a hash is that the check fails if there is even the slightest numerical difference.  And we find such differences across Matlab versions and computer hardware.

You might think you could round and then hash, and although that might be true, our attempts at this in the UnitTestToolbox never worked.  We’d think we had stuff robust, and then a hash-based check would fail because of some small numerical difference that was just larger than whatever tolerance we set.

Coded the test in terms of checking that a sum of values rounded to 5 places remains the same.  I think that will catch any problems and be more robust.  Also allows checking size of failure if one pops up later, which saving the has does not.

Left the hash code in the example commented out, in case someone wants to check it on the hardware/version it was set up for.  But recommend deleting once that check is done.