Bug fix to added noise
The colored noise returned by dsp.ColoredNoise can have a substantial non-zero component. While understandable, this can be a problem for simulations that assume that the model response is zero centered. This bug fix mean-centers the noise prior to scaling to the desired SD, and being added to the model fit.

Passes IAMP validations.